refactor: requestes changes
Moved class to its own file; created private fact member; added
devcontainer extensions
