+++ .vscode/settings.json
@@ -7,17 +7,24 @@
     "Pharma",
     "TSLA",
     "WLRH",
+    "adapter",
     "corejs",
     "csst",
+    "esbenp",
     "gaap",
     "gryffindor",
+    "hbenl",
     "hufflepuf",
+    "mocha",
     "naruto",
     "newmsft",
     "numcommadecimal",
     "numdotdecimal",
+    "prettier",
     "ravenclawn",
     "slytherin",
+    "test",
+    "vscode",
     "xbrl",
     "xbrldi",
     "xbrli"
